Частота tone в два раза ниже

+++ Syntez/pins.h
@@ -68,6 +68,6 @@ class OutputPCF8574 {
 	  void Write();
 };
 
-void OutputTone(uint8_t pin, uint8_t value);
+void OutputTone(uint8_t pin, int value);
 
 #endif
